Customizable extention, 404 page, check URL for menu item
There are two new options at the top of the file to allow the user
to specifiy what file extension they wish to use as well as what
404 page they wish to use. The menu items are checked to make sure
that have a non-empty URL. Previously a file could be created with
a TITLE and a blank URL breaking the anchor tag that would be
generated.

I also cleaned up the <?php echos ?> by using the <?= syntax which
is cleaner and more readable.